Small things changed to buttons

+++ backend/Domain/Classes/Review.cs
@@ -5,7 +5,7 @@ public class Review
     public Guid Id { get; private set; }
     public User Creator { get; private set; }
     public string Title { get; private set; }
-    public int Rating { get; private set; }
+    public double Rating { get; private set; }
     public string Description { get; private set; }
     public ItemType ItemType { get; private set; }
     public DateTime LastUpdated { get; private set; }
@@ -14,7 +14,7 @@ public class Review
     
     public IReadOnlyCollection<Group> Groups => _groups.AsReadOnly();
     
-    public Review(User creator, string title, int rating, string description, ItemType itemType, List<Group> groups)
+    public Review(User creator, string title, double rating, string description, ItemType itemType, List<Group> groups)
     {
         Id = Guid.NewGuid();
         Creator = creator;
@@ -29,7 +29,7 @@ public class Review
             _groups.AddRange(groups);
         }
     }
-    public Review(User creator, string title, int rating, string description, ItemType itemType) 
+    public Review(User creator, string title, double rating, string description, ItemType itemType) 
         : this(creator, title, rating, description, itemType, new List<Group>())
     {
     }
@@ -80,7 +80,7 @@ public class Review
         ChangeLastUpdated();
     }
 
-    public void ChangeRating(int rating)
+    public void ChangeRating(double rating)
     {
         Rating = rating;
         ChangeLastUpdated();
